+++ packages/terra-time-input/tests/nightwatch/TimeInputTestRoutes.jsx
@@ -5,6 +5,7 @@ import { Route } from 'react-router';
 import TimeInputTests from './TimeInputTests';
 import TimeInputDefault from './component/TimeInputDefault';
 import TimeInputDefaultTime from './component/TimeInputDefaultTime';
+import TimeInputDisabled from './component/TimeInputDisabled';
 import TimeInputOnChange from './component/TimeInputOnChange';
 import TimeInputInvalidDate from './component/TimeInputInvalidDate';
 import TimeInputTwelveHour from './component/TimeInputTwelveHour';
@@ -17,6 +18,7 @@ const routes = (
     <Route path="/tests/time-input-tests" component={TimeInputTests} />
     <Route path="/tests/time-input-tests/default" component={TimeInputDefault} />
     <Route path="/tests/time-input-tests/default-time" component={TimeInputDefaultTime} />
+    <Route path="/tests/time-input-tests/disabled" component={TimeInputDisabled} />
     <Route path="/tests/time-input-tests/on-change" component={TimeInputOnChange} />
     <Route path="/tests/time-input-tests/invalid-time" component={TimeInputInvalidDate} />
     <Route path="/tests/time-input-tests/twelve-hour" component={TimeInputTwelveHour} />