+++ packages/terra-alert/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Update to React 16
+* Uplift to Button 2.0
 
 1.16.0 - (February 1, 2018)
 ------------------