+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,8 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
-### Changed
+### Added
 * Update to React 16
+*Added disabled prop.
+
+### Changed
+* Uplift to Button 2.0
 
 1.7.0 - (February 1, 2018)
 ------------------