+++ packages/terra-menu/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Update to React 16
+* Changed React.Children.map to React.Children.forEach to prevent '.$' being prepended to the key.
 
 1.13.0 - (February 1, 2018)
 ------------------