+++ packages/terra-site/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 ### Changed
 * Update to Reat 16
 * Removed unknown `isDisabled` prop from InlineRadiosExample.jsx for terra-form-radio.
+* Update Collapsible Menu View and Toggle Button examples to use Button 2.0.
 ### Removed
 * terra-theme-provider examples
 * terra-app-delegate examples