+++ packages/terra-status-view/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ------------------
 ### Changed
 * Update to React 16
+* Update to Button 2.0
 
 1.1.0 - (February 1, 2018)
 ------------------