+++ packages/terra-time-input/CHANGELOG.md
@@ -3,8 +3,12 @@ Changelog
 
 Unreleased
 ----------
+### Added
+*Added disabled prop.
+
 ### Changed
 * Update to React 16
+* Uplift to ButtonGroup 2.0
 
 1.18.0 - (February 1, 2018)
 ------------------